backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aioodbc==0.5.0
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
deprecation==2.1.0
fastapi==0.115.0
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gotrue==2.9.1
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
postgrest==0.17.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyodbc==5.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
storage3==0.8.0
StrEnum==0.4.15
supabase==2.8.1
supafunc==0.6.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
websockets==13.1
yarl==1.13.1
